+++ packages/services/src/state_listener.rs
@@ -65,7 +65,7 @@ where
 
                 match (tx.state, self.l1_adapter.is_squeezed_out(tx.hash).await?) {
                     (TransactionState::Pending, true) => {
-                        //not in the mempool anymore set it to failed
+                        // not in the mempool anymore set it to failed
                         selective_change.push((tx.hash, tx.nonce, TransactionState::Failed));
 
                         info!(
