--- packages/utils/src/old.rs
@@ -1,3 +1,0 @@
-pub mod copied_from_alloy;
-pub mod encoder;
-pub mod new_encoder;
