+++ packages/utils/tests/public_api.rs
@@ -1,17 +1,15 @@
 #[cfg(test)]
 mod test {
-
-    use alloy::{
-        consensus::{BlobTransactionSidecar, EnvKzgSettings},
-        eips::eip4844::DATA_GAS_PER_BLOB,
-    };
-    use c_kzg::{KzgCommitment, KzgProof};
+    use alloy::{consensus::EnvKzgSettings, eips::eip4844::DATA_GAS_PER_BLOB};
     use itertools::Itertools;
     use proptest::prelude::*;
     use rand::{rngs::SmallRng, seq::SliceRandom, RngCore, SeedableRng};
     use test_case::test_case;
-    use utils::{
-        decoder::NewDecoder, encoder::NewEncoder, generate_sidecar, Blob, BlobHeader, BlobHeaderV1,
+    use utils::blob::{
+        decoder::NewDecoder,
+        encoder::NewEncoder,
+        generate_sidecar,
+        header::{BlobHeader, BlobHeaderV1},
     };
 
     #[test_case(1,  1; "one blob")]
@@ -21,7 +19,7 @@ mod test {
     #[test_case(130037 * 2  + 1,  3; "three blobs")]
     fn can_calculate_blobs_needed_without_encoding(num_bytes: usize, num_blobs: usize) {
         // given
-        let encoder = NewEncoder {};
+        let encoder = NewEncoder::default();
 
         // when
         let usage = encoder.blobs_needed_to_encode(num_bytes);
@@ -36,7 +34,7 @@ mod test {
         #[test]
         fn calculated_blobs_needed_the_same_as_if_you_encode(byte_amount in 1..=DATA_GAS_PER_BLOB*20) {
             // given
-            let encoder = NewEncoder{};
+            let encoder = NewEncoder::default();
 
             // when
             let usage = encoder.blobs_needed_to_encode(byte_amount as usize);
@@ -48,7 +46,7 @@ mod test {
         #[test]
         fn full(byte_amount in 1..=DATA_GAS_PER_BLOB*20) {
         // given
-        let encoder = NewEncoder {};
+        let encoder = NewEncoder::default();
 
         let mut data = vec![0; byte_amount as usize];
         let mut rng = SmallRng::from_seed([0; 32]);
@@ -72,7 +70,7 @@ mod test {
         }
 
         // can be decoded into original data
-        let decoded_data  = NewDecoder{}.decode(&blobs).unwrap();
+        let decoded_data  = NewDecoder::default().decode(&blobs).unwrap();
         proptest::prop_assert_eq!(data, decoded_data);
         }
     }
@@ -84,7 +82,7 @@ mod test {
     #[test_case(1_200_000)]
     fn can_generate_proofs_and_committments_for_encoded_blobs(byte_num: usize) {
         // given
-        let encoder = NewEncoder {};
+        let encoder = NewEncoder::default();
 
         let mut data = vec![0; byte_num];
         let mut rng = SmallRng::from_seed([0; 32]);
@@ -110,14 +108,14 @@ mod test {
     #[test_case(1_200_000)]
     fn blobs_can_be_decoded_when_in_order(byte_num: usize) {
         // given
-        let encoder = NewEncoder {};
+        let encoder = NewEncoder::default();
 
         let mut data = vec![0; byte_num];
         let mut rng = SmallRng::from_seed([0; 32]);
         rng.fill_bytes(&mut data[..]);
         let blobs = encoder.encode(&data, 0).unwrap();
 
-        let decoder = NewDecoder {};
+        let decoder = NewDecoder::default();
 
         // when
         let decoded_data = decoder.decode(&blobs).unwrap();
@@ -132,7 +130,7 @@ mod test {
     #[test_case(1_200_000)]
     fn blobs_can_be_decoded_even_if_shuffled_around(byte_num: usize) {
         // given
-        let encoder = NewEncoder {};
+        let encoder = NewEncoder::default();
 
         let mut rng = SmallRng::from_seed([0; 32]);
         let data = {
@@ -147,7 +145,7 @@ mod test {
             blobs
         };
 
-        let decoder = NewDecoder {};
+        let decoder = NewDecoder::default();
 
         // when
         let decoded_data = decoder.decode(&blobs).unwrap();
@@ -162,7 +160,7 @@ mod test {
     fn roundtrip_header_encoding(num_bytes: usize, bundle_id: u32) {
         // given
         let blob = {
-            let encoder = NewEncoder {};
+            let encoder = NewEncoder::default();
             encoder
                 .encode(&vec![0; num_bytes], bundle_id)
                 .unwrap()
@@ -170,7 +168,7 @@ mod test {
                 .unwrap()
         };
 
-        let decoder = NewDecoder {};
+        let decoder = NewDecoder::default();
 
         // when
         let header = decoder.read_header(&blob).unwrap();
